+++ packages/bridge/bridge-react/src/utils.ts
@@ -1,4 +1,5 @@
 import React from 'react';
+import { FederationHost } from '@module-federation/runtime';
 import { createLogger } from '@module-federation/sdk';
 
 export const LoggerInstance = createLogger(
@@ -41,3 +42,24 @@ export function pathJoin(...args: string[]) {
   }, '');
   return res || '/';
 }
+
+export const getModuleName = (id: string) => {
+  if (!id) {
+    return id;
+  }
+  // separate module name without detailed module path
+  // @vmok-e2e/edenx-demo-app2/button -> @vmok-e2e/edenx-demo-app2
+  const idArray = id.split('/');
+  if (idArray.length < 2) {
+    return id;
+  }
+  return idArray[0] + '/' + idArray[1];
+};
+
+export const getRootDomDefaultClassName = (moduleName: string) => {
+  if (!moduleName) {
+    return '';
+  }
+  const name = getModuleName(moduleName).replace(/\@/, '').replace(/\//, '-');
+  return `bridge-root-component-${name}`;
+};