+++ packages/bridge/vue3-bridge/src/remoteApp.tsx
@@ -1,7 +1,8 @@
 import { ref, onMounted, onBeforeUnmount, watch, defineComponent } from 'vue';
+import { dispatchPopstateEnv } from '@module-federation/bridge-shared';
 import { useRoute } from 'vue-router';
 import { LoggerInstance } from './utils';
-import { dispatchPopstateEnv } from '@module-federation/bridge-shared';
+import { getInstance } from '@module-federation/runtime';
 
 export default defineComponent({
   name: 'RemoteApp',
@@ -16,11 +17,13 @@ export default defineComponent({
     const providerInfoRef = ref(null);
     const pathname = ref('');
     const route = useRoute();
+    const hostInstance = getInstance();
 
     const renderComponent = () => {
       const providerReturn = props.providerInfo?.();
       providerInfoRef.value = providerReturn;
-      const renderProps = {
+
+      let renderProps = {
         name: props.moduleName,
         dom: rootRef.value,
         basename: props.basename,
@@ -30,7 +33,15 @@ export default defineComponent({
         `createRemoteComponent LazyComponent render >>>`,
         renderProps,
       );
+
+      const beforeBridgeRenderRes =
+        hostInstance?.bridgeHook?.lifecycle?.beforeBridgeRender?.emit(
+          renderProps,
+        ) || {};
+
+      renderProps = { ...renderProps, ...beforeBridgeRenderRes.extraProps };
       providerReturn.render(renderProps);
+      hostInstance?.bridgeHook?.lifecycle?.afterBridgeRender?.emit(renderProps);
     };
 
     const watchStopHandle = watch(
@@ -61,7 +72,21 @@ export default defineComponent({
         ...props,
       });
       watchStopHandle();
+
+      hostInstance?.bridgeHook?.lifecycle?.beforeBridgeDestroy?.emit({
+        name: props.moduleName,
+        dom: rootRef.value,
+        basename: props.basename,
+        memoryRoute: props.memoryRoute,
+      });
+
       (providerInfoRef.value as any)?.destroy({ dom: rootRef.value });
+      hostInstance?.bridgeHook?.lifecycle?.afterBridgeDestroy?.emit({
+        name: props.moduleName,
+        dom: rootRef.value,
+        basename: props.basename,
+        memoryRoute: props.memoryRoute,
+      });
     });
 
     return () => <div ref={rootRef}></div>;