+++ packages/node/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/node
 
+## 2.6.9
+
+### Patch Changes
+
+- Updated dependencies [ff8ce29]
+  - @module-federation/runtime@0.7.4
+  - @module-federation/enhanced@0.7.4
+  - @module-federation/sdk@0.7.4
+  - @module-federation/utilities@3.1.27
+
 ## 2.6.8
 
 ### Patch Changes