+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @module-federation/rsbuild-plugin
 
+## 0.7.4
+
+### Patch Changes
+
+- @module-federation/enhanced@0.7.4
+- @module-federation/sdk@0.7.4
+
 ## 0.7.3
 
 ### Patch Changes