+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @module-federation/sdk
 
+## 0.7.4
+
 ## 0.7.3
 
 ### Patch Changes